+++ CHANGES.md
@@ -17,6 +17,7 @@ Roadmap
 * Enhancement of combination CLI utility with multiple files output.
 * Enhancement of flash CLI utility with multiple files output.
 * Filter CLI selection mode based on count of passed filters.
+* Adding difference CLI utility.
 * Removal of 99 balls limitation in DrawBox class.
 * Package fix.
 * Documentation update.